Feture/remove ttl expired (#139)
* Added feature to index/remove ttl expiring cells to index.

* Added numDocs checks in TTL remove testsAT

* Moved ttl tests to separate suite, renamed to *AT, rolled back some pointless edits, added printStackTrace to JMX errors, replaced for loops, used String.format in JMX

* Refactored CassandraUtils.doWithEachJmxService, added TTL doc in CHANGELOG and documentation

+++ plugin/src/main/java/com/stratio/cassandra/lucene/IndexWriterSkinny.java
@@ -66,18 +66,20 @@ class IndexWriterSkinny extends IndexWriter {
     /** {@inheritDoc} */
     @Override
     public void finish() {
-        optionalRow.ifPresent(row -> {
-            if (service.needsReadBeforeWrite(key, row)) {
-                UnfilteredRowIterator iterator = service.read(key, nowInSec, opGroup);
-                if (iterator.hasNext()) {
-                    row = (Row) iterator.next();
+        if (transactionType != IndexTransaction.Type.CLEANUP) {
+            optionalRow.ifPresent(row -> {
+                if (service.needsReadBeforeWrite(key, row) || (transactionType == IndexTransaction.Type.COMPACTION)) {
+                    UnfilteredRowIterator iterator = service.read(key, nowInSec, opGroup);
+                    if (iterator.hasNext()) {
+                        row = (Row) iterator.next();
+                    }
                 }
-            }
-            if (row.hasLiveData(nowInSec)) {
-                service.upsert(key, row, nowInSec);
-            } else {
-                service.delete(key);
-            }
-        });
+                if (row.hasLiveData(nowInSec)) {
+                    service.upsert(key, row, nowInSec);
+                } else {
+                    service.delete(key);
+                }
+            });
+        }
     }
 }
